+++ src/shared/domain/entities/index.ts
@@ -1,1 +1,4 @@
+export * from './author';
+export * from './category';
+export * from './featuredImage';
 export * from './post';
