+++ src/shared/domain/useCases/fetchPost.ts
@@ -1,11 +1,6 @@
 import { Author, Category, IPostRepository, Post } from '@/domain';
 import { PostRepository } from '@/infrastructure/post';
 
-// export const fetchPost = async (slug: string): Promise<PostResponse> => {
-//   const response = await getPost(slug);
-//   return response;
-// };
-
 export class FetchPostUseCase {
   private readonly postRepository: IPostRepository;
 
@@ -26,7 +21,10 @@ export class FetchPostUseCase {
       excerpt: post.excerpt,
       content: post.content,
       date: post.date,
-      featuredImageUrl: post.featuredImage.node.sourceUrl ?? '/static/images/not_found.png',
+      featuredImageUrl: {
+        url: post.featuredImage.node.sourceUrl ?? '/static/images/not_found.png',
+        alt: post.featuredImage.node.altText
+      },
       author,
       categories
     };
