+++ src/shared/domain/useCases/fetchPosts.ts
@@ -3,21 +3,6 @@ import { isDevelopment } from '@/extension';
 import { PostRepository } from '@/infrastructure/post';
 import { Node, PostsResponse } from '@/infrastructure/types';
 
-// export const fetchAllPosts = async (
-//   posts: Array<Node>,
-//   _offset: string
-// ): Promise<{ nodes: Array<Node>; hasNextPage: boolean; offset: string }> => {
-//   const res: PostsResponse = await getAllPosts(100, _offset);
-//   if (!res.data.posts.pageInfo.hasNextPage || isDevelopment()) {
-//     return {
-//       nodes: [...posts, ...res.data.posts.edges],
-//       hasNextPage: res.data.posts.pageInfo.hasNextPage,
-//       offset: res.data.posts.pageInfo.endCursor
-//     };
-//   }
-//   return fetchAllPosts([...posts, ...res.data.posts.edges], res.data.posts.pageInfo.endCursor);
-// };
-
 export class FetchPostsUseCase {
   private readonly postRepository: IPostRepository;
 
@@ -53,7 +38,10 @@ export class FetchPostsUseCase {
         excerpt: post.excerpt,
         content: post.content,
         date: post.date,
-        featuredImageUrl: post.featuredImage.node.sourceUrl ?? '/static/images/not_found.png',
+        featuredImageUrl: {
+          url: post.featuredImage.node.sourceUrl ?? '/static/images/not_found.png',
+          alt: post.featuredImage.node.altText
+        },
         author,
         categories
       };
